requirements.txt
Django==3.2.6
wagtail==2.14.1
wagtail-localize==1.1.0
django-countries==7.2.1
django-allauth==0.45.0
django-widget-tweaks==1.4.8
django-extensions==3.1.3
django-contrib-comments==2.1.0
django-modeltranslation==0.17.3